fix(engine): coding ops honor the documented memory budget
_unpack_symbols built an int64 (n, width) matrix - measured 9x the input,
vs the 2-3x the bitfile guard promised, so the guard admitted captures it
could not process. Column-loop int32 unpack/pack (~1x), uint8 XOR-reduce
syndrome in block_decode (10.7x -> ~7x), float32 m_slice; budget comment
now states the measured multipliers.

+++ src/marconi/engine/coding/ops_symbols.py
@@ -33,7 +33,7 @@ def m_slice_rx(
     if sym is None:
         return c
     regions = np.searchsorted(
-        np.asarray(thresholds, np.float64), np.asarray(sym, np.float64)
+        np.asarray(thresholds, np.float32), np.asarray(sym, np.float32)
     )
     mapped = np.asarray(levels, np.int16)[regions]
     return replace(c, symbols=mapped.astype(np.int16))
